Fix all tests

+++ test/misc/unit/libraries/BytesLib.t.sol
@@ -54,7 +54,7 @@ contract BytesLibTest is Test {
         assertEq(BytesLib.toUint64(value, randomStart.length), number);
     }
 
-    function testToUint88(uint64 number, bytes memory randomStart, bytes memory randomEnd) public pure {
+    function testToUint88(uint88 number, bytes memory randomStart, bytes memory randomEnd) public pure {
         bytes memory value = bytes.concat(bytes.concat(randomStart, abi.encodePacked(number)), randomEnd);
         assertEq(BytesLib.toUint88(value, randomStart.length), number);
     }
